Implement power operator (**) for f32 and add corresponding tests for int and float types
