+++ spy/vm/modules/operator/binop.py
@@ -144,6 +144,7 @@ MM.register("*",  "f32", "f32", OP.w_f32_mul)
 MM.register("/",  "f32", "f32", OP.w_f32_div)
 MM.register("//", "f32", "f32", OP.w_f32_floordiv)
 MM.register("%",  "f32", "f32", OP.w_f32_mod)
+MM.register("**", "f32", "f32", OP.w_f32_pow)
 MM.register("==", "f32", "f32", OP.w_f32_eq)
 MM.register("!=", "f32", "f32", OP.w_f32_ne)
 MM.register("<" , "f32", "f32", OP.w_f32_lt)
@@ -160,6 +161,8 @@ for int_t in ["i32"]:
     MM.register("*",  int_t, "f32", OP.w_f32_mul)
     MM.register("/",  "f32", int_t, OP.w_f32_div)
     MM.register("/",  int_t, "f32", OP.w_f32_div)
+    MM.register("**", "f32", int_t, OP.w_f32_pow)
+    MM.register("**", int_t, "f32", OP.w_f32_pow)
     MM.register("==", "f32", int_t, OP.w_f32_eq)
     MM.register("==", int_t, "f32", OP.w_f32_eq)
     MM.register("!=", "f32", int_t, OP.w_f32_ne)
